+++ src/components/ui/my-table/index.ts
@@ -1,0 +1,3 @@
+export { MyTable } from './my-table'
+export type { MyTableColumn, MyTableColumnAlign, MyTableProps } from './type'
+export { resolveColumnLayoutStyle } from './utils'